portweb
This is an HTML code for a personal portfolio website that showcases the student's personal and educational background, as well as their projects. The website has a dark background color and a clean layout, with separate sections for HOME, SKILLS, ABOUT, and PROJECT, and social icons fixed to the left side of the screen.